pyproject.toml
[tool.poetry]
name = "human-recognition"
version = "0.1.0"
description = "Projeto de reconhecimento de silhueta humana com LBP e Random Forest"
authors = ["Seu Nome <cavalcanteprofissional@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.10"
opencv-python = "^4.8.1"
numpy = "^1.24.0"
scikit-learn = "^1.3.0"
scikit-image = "^0.22.0"
python-dotenv = "^1.0.0"
kaggle = "^1.5.16"
matplotlib = "^3.7.0"
seaborn = "^0.13.0"
joblib = "^1.3.0"
pillow = "^10.0.0"
pandas = "^2.0.0"
streamlit = "^1.28.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.0.0"
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"